+++ .gitignore
@@ -26,4 +26,4 @@ vite.config.ts.timestamp-*
 # ESLint cache
 .eslintcache
 
-static/images/blog/*-w*.webp
+static/images/**/*-w*.webp
