+++ package.json
@@ -1,7 +1,7 @@
 {
 	"name": "joshuafolkken-com",
 	"private": true,
-	"version": "0.63.4",
+	"version": "0.64.0",
 	"type": "module",
 	"scripts": {
 		"setup:security": "safe-chain setup",
